requirements.txt
streamlit==1.24.0
ultralytics==8.0.146
torch==2.2.0
torchvision==0.17.0
numpy
opencv-python
Pillow
matplotlib